feat: implement multi-unit support and decimal quantities for inventory and sales
- Add `UnitType` enum to support discrete, weight, volume, and length-based products.
- Refactor stock and cart quantity logic from `Int` to `Double` to support fractional units.
- Introduce `QuantityInput` component for specialized decimal and discrete quantity entry.
- Integrate unit selection and formatting across Add Product, Inventory, Checkout, and Receipt screens.
- Enhance `SessionManager` to track and manage active branch selection for admins and staff.
- Update network DTOs to persist unit type information.
- Upgrade Gradle wrapper to 9.5.1 and project JDK to jbr-25.
- Add suffix support to `CustomTextField`.

+++ app/src/main/java/com/techsultan/zenithpro/features/inventory/data/local/ProductStockEntity.kt
@@ -24,8 +24,8 @@ import com.techsultan.zenithpro.core.util.Util
 data class ProductStockEntity(
     @PrimaryKey val id: String,
     val variantId: String,
-    val branchId: String,       // ADD
-    val quantity: Int,
+    val branchId: String,
+    val quantity: Double,
     val expiryDate: String?,
     val lowStockAlert: Int?,
     val updatedAt: String,
